Fixing Pneumatics + Speed Algorhythm
There was an issue where all controls would reset the arm to its non-deployed position.  This was fixed  by removing the command that prevented the arm from being out and in and the same time (which was irrelevent since this was all controlled in 1 class).  The speed algorhythm was made to be cubic so that it would have a sensitivity curve and be easier to move at slow speeds